ルート結果を表示するページのApp Routerのルーティングを修正
